src/com/lgm/net/TankMsgDecoder.java
package com.lgm.net;

import io.netty.buffer.ByteBuf;
import io.netty.channel.ChannelHandlerContext;
import io.netty.handler.codec.ByteToMessageDecoder;

import java.util.List;

/**
 * @author:李罡毛
 * @date:2021/3/29 15:59
 */
public class TankMsgDecoder extends ByteToMessageDecoder {
    @Override
    protected void decode(ChannelHandlerContext channelHandlerContext, ByteBuf byteBuf, List<Object> list) throws Exception {
        if (byteBuf.readableBytes() < 8) return;
        list.add(new TankMsg(byteBuf.readInt(),byteBuf.readInt()));
    }
}
